Move modals related object into lazy loads chunk
