+++ src/ggrc_workflows/assets/javascripts/components/workflow_clone.js
@@ -5,30 +5,53 @@
 
 import {BUTTON_VIEW_SAVE_CANCEL} from '../../../../ggrc/assets/javascripts/plugins/utils/modals';
 
-(function (GGRC, can) {
-  'use strict';
+let CloneWorkflow = can.Model.Cacheable({
+  defaults : {
+    clone_people: true,
+    clone_tasks: true,
+    clone_objects: true
+  }
+}, {
+  refresh: function() {
+    return $.when(this);
+  },
+  save: function() {
+    var workflow = new CMS.Models.Workflow({
+      clone: this.source_workflow.id,
+      context: null,
+      clone_people: this.clone_people,
+      clone_tasks: this.clone_tasks,
+      clone_objects: this.clone_objects
+    });
 
-  can.Component.extend({
-    tag: "workflow-clone",
-    template: "<content/>",
-    events: {
-      click: function(el) {
-        var workflow, $target;
+    return workflow.save().then(function(workflow) {
+      GGRC.navigate(workflow.viewLink);
+      return this;
+    });
 
-        $target = $('<div class="modal hide"></div>').uniqueId();
-        $target.modal_form({}, el);
-        import(/* webpackChunkName: "modalsCtrls" */'../../../../ggrc/assets/javascripts/controllers/modals')
-          .then(() => {
-            $target.ggrc_controllers_modals({
-              modal_title: "Clone Workflow",
-              model: CMS.ModelHelpers.CloneWorkflow,
-              instance: new CMS.ModelHelpers.CloneWorkflow({source_workflow: this.scope.workflow}),
-              content_view: GGRC.mustache_path + "/workflows/clone_modal_content.mustache",
-              custom_save_button_text: "Proceed",
-              button_view: BUTTON_VIEW_SAVE_CANCEL
-            });
+  }
+});
+
+can.Component.extend({
+  tag: "workflow-clone",
+  template: "<content/>",
+  events: {
+    click: function(el) {
+      var $target;
+
+      $target = $('<div class="modal hide"></div>').uniqueId();
+      $target.modal_form({}, el);
+      import(/* webpackChunkName: "modalsCtrls" */'../../../../ggrc/assets/javascripts/controllers/modals')
+        .then(() => {
+          $target.ggrc_controllers_modals({
+            modal_title: "Clone Workflow",
+            model: CloneWorkflow,
+            instance: new CloneWorkflow({source_workflow: this.scope.workflow}),
+            content_view: GGRC.mustache_path + "/workflows/clone_modal_content.mustache",
+            custom_save_button_text: "Proceed",
+            button_view: BUTTON_VIEW_SAVE_CANCEL
           });
-      }
+        });
     }
-  });
-})(window.GGRC, window.can);
+  }
+});
